+++ build.rs
@@ -6,7 +6,7 @@ fn main() {
     protobuf_codegen::Codegen::new()
         .pure()
         .out_dir(out_dir)
-        .inputs(["protos/rendezvous.proto", "protos/message.proto"])
+        .inputs(["protos/rendezvous.proto"])
         .include("protos")
         .customize(protobuf_codegen::Customize::default().tokio_bytes(true))
         .run()
